Move tests to expected package
